Add deprecation note  for notifying visitors
Hi folks, Rapi here from the CTO's office. NYC PPE Dashboard's starting the process of deprecating this and we're hoping to deprecate it by October 31. Would it be possible to add this to the code as well as the Heroku instance? I understand it's currently running in Heroku? Would love to chat and touch base! Thanks!